checkpoint(library/Byte): renames core logic to "Byte.ts"
