+++ src/library/Byte.ts
