+++ src/library/Byte/exports.ts
@@ -1,3 +1,3 @@
 export {
   Byte_cofactorTo as cofactorTo,
-} from './utilities/cofactorTo';
+} from '../Byte.ts';
